Move restitution solver into separate file.
